End of day Thursday

+++ Models/Item.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public int CategoryId { get; set; }
-        public float Weight { get; set; }
+        public double Weight { get; set; }
         public int FoodTypeId { get; set; }
         public string Image { get; set; }
         public int Quantity { get; set; }
